+++ iree/compiler/Codegen/LLVMGPU/Passes.cpp
@@ -6,7 +6,7 @@
 
 #include "iree/compiler/Codegen/Passes.h"
 
-#include "iree-dialects/Dialect/LinalgExt/Transforms/Passes.h"
+#include "iree-dialects/Dialect/LinalgExt/Passes/Passes.h"
 #include "iree/compiler/Codegen/PassDetail.h"
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
 #include "mlir/Conversion/GPUToNVVM/GPUToNVVMPass.h"
