+++ iree/compiler/Codegen/SPIRV/Passes.cpp
@@ -13,7 +13,7 @@
 
 #include "iree/compiler/Codegen/Passes.h"
 
-#include "iree-dialects/Dialect/LinalgExt/Transforms/Passes.h"
+#include "iree-dialects/Dialect/LinalgExt/Passes/Passes.h"
 #include "iree/compiler/Codegen/PassDetail.h"
 #include "iree/compiler/Codegen/Passes.h"
 #include "iree/compiler/Codegen/SPIRV/MemorySpace.h"
